Config/Magyarnarancs_v0
The commit contains the config files for 'magyarnarancs', a version of the XML that contains placeholder values for required attributes, and a line for the portal in the warc2portalname.yaml.
Produces an unresolved error "The tag is not in the dictionary.The dictionary needs to be updated" for the two 'div' tags that contain the article text.